Use usize instead of i32 for future portability

+++ assembly/sum.ts
@@ -1,6 +1,6 @@
-export function sum(from: i32, to: i32): f64 {
+export function sum(from: usize, to: usize): f64 {
   let total: f64 = 0;
-  for (let i: i32 = from; i < to; i += 1) {
+  for (let i: usize = from; i < to; i += 1) {
     total += load<f64>(i << 3);
   }
   return total;
